Move KubernetesClient to k8sutil package.
